+++ src/components/Wishlist/Wishlist.js
@@ -1,8 +1,8 @@
 import { useContext } from "react";
-import "./wishlist.css";
 
 import ProductList from "../ProductListing/ProductList/ProductList";
 import { WishlistContext } from "../../store/wishlist/wishlistContext";
+import "./wishlist.css";
 
 export default function Wishlist() {
   const { wishlist } = useContext(WishlistContext);
